+++ FE/src/App.jsx
@@ -517,7 +517,6 @@ const startJobPolling = (jobId) => {
         return;
       }
 
-      addSystem("시뮬레이션 시작 (실시간 로그 수집)");
       startJobPolling(kick.job_id);
     } catch (err) {
       console.error("시뮬레이션 실행 실패:", err);
